+++ src/main/java/com/jih/dto/CategoryDto.java
@@ -1,0 +1,12 @@
+package com.jih.dto;
+
+
+import java.time.LocalDateTime;
+
+public record CategoryDto(
+        Long id,
+        String name,
+        LocalDateTime createdAt,
+        LocalDateTime updatedAt
+) {
+}