+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Features:
  - YP - Add PostgreSQL persistence with JPA and Liquibase
  - YP - Add Question controller and service tests
  - YP - Add Question UUID, and Category entity
+ - YP - Add Category controller and service
 
 ## jih-app 0.2.0